Introduced an experimental query endpoint.
Some short notes on the oddities and features of such a limited
Wikidata Query Endpoint.

## 1
Wikidata propertyIds are here called "Association Types" and any of such
_assocTypes_ (for short) can have _one_ or _more_ propertyIds assigned.
This allows administrators to interactively configure an API behaviour
similar to the one of calculating "type trees" or "synonyms" over many
properties but in foresight (while drinking a cup of coffee) and not at
runtime. As new properties come by (resp. are created in other languages)
these can simple be matched on all existing assocations (in the DB) by the
means of "increasing" or "decreasing" the semantics of an Association Type
known to this endpoint by propertyIds.

Managing these assignments can be done in Topicmaps by using the standard dm4-webclient.

## 2
Claims between items, once imported, are always treated as bi-directional.
This means, once a relation was claimed at one of the items on wikidata.org
this endpoint behaves as if the claim (identified by a given propertyId) was
set on both items.

## 3
The import of wikidata entities and associations between those currently relies
on a hardcoded mapping of wikidata entities to four core types of the
DeepaMehta 4 Standard Distribution.

The following core types are imported in the case of an _instanceOf_ || _subclassOf_ claim:
  _Persons_ 	 	== HUMAN (Q5) or PERSON (Q215627)
  _Institutions_ 	== COMPANY (Q783794), UNIVERSITY (Q3918),
                           ORGANISATION (Q43229), COLLEGIATE UNIVERSITY (Q3354859)
  _Cities_ 		== CITY (Q515), METROPOLIS (Q200250), CAPITAL_CITY (Q36)
  _Countries_ 	 	== COUNTRY (Q6256), SOUVEREIGN_STATE (Q3624078)

The following associations are created between all those four core items (if claims with
the following propertyIds were made):
  affiliated_with	== isMemberOf (P463) || isPartyMemberOf (P102) ||
   			   isAffiliatedWith (P1416) || isAlmaMaterOf (P69)
  employee_of  		== isEmployeeOf (P108)
  citizen_of		== isCitizenOf (P27) || isOfficialyResidingAt (P263)
  student_of		== isStudentOf (P1066) || isDoctoralStudenfOf (P185)
  mentor_of		== isDoctoralAdvisorOf (P184)

Two examplary GET Request to this API might be:

- `/wdtk/query/Q42/P108` 	>>> BBC (Q9531)
- `/wdtk/query/Q9531/P108` 	>>> Douglas Adams (and three others ..)

While a list of publicly known employees of BBC (in wikidata) is most definitely not to be stored on the BBC item entity, this info may be stored on the respective persons item entity (see point 2 above).